code/requirements.txt
jsonlines>=1.2.0,<2.0.0
praw>=7.1.0,<8.0.0
progressbar2>=3.51.4,<4.0.0
typeguard>=2.7.1,<3.0.0
